+++ packages/core/src/casbin-secure-sequence.ts
@@ -1,5 +1,6 @@
 import {inject} from '@loopback/context';
 import {
+  ExpressRequestHandler,
   FindRoute,
   HttpErrors,
   InvokeMethod,
@@ -49,6 +50,8 @@ export class CasbinSecureSequence implements SequenceHandler {
    */
   @inject(SequenceActions.INVOKE_MIDDLEWARE, {optional: true})
   protected invokeMiddleware: InvokeMiddleware = () => false;
+  @inject(SFCoreBindings.EXPRESS_MIDDLEWARES, {optional: true})
+  protected expressMiddlewares: ExpressRequestHandler[] = [];
 
   constructor(
     // sonarignore:start
@@ -95,6 +98,15 @@ export class CasbinSecureSequence implements SequenceHandler {
         Remote Address = ${request.connection.remoteAddress}
         Remote Address (Proxy) = ${request.headers['x-forwarded-for']}`,
       );
+
+      if (this.expressMiddlewares?.length) {
+        const responseGenerated = await this.invokeMiddleware(
+          context,
+          this.expressMiddlewares,
+        );
+        if (responseGenerated) return;
+      }
+
       const finished = await this.invokeMiddleware(context);
       if (finished) return;
       const route = this.findRoute(request);
@@ -111,6 +123,7 @@ export class CasbinSecureSequence implements SequenceHandler {
       await this.authenticateClientRequest(request);
       const authUser: IAuthUserWithPermissions = await this.authenticateRequest(
         request,
+        response,
       );
       const resVal = await this.casbinResModifierFn(args, request);