+++ packages/core/src/secure-sequence.ts
@@ -1,5 +1,6 @@
 import {inject} from '@loopback/context';
 import {
+  ExpressRequestHandler,
   FindRoute,
   HttpErrors,
   InvokeMethod,
@@ -48,6 +49,8 @@ export class SecureSequence implements SequenceHandler {
    */
   @inject(SequenceActions.INVOKE_MIDDLEWARE, {optional: true})
   protected invokeMiddleware: InvokeMiddleware = () => false;
+  @inject(SFCoreBindings.EXPRESS_MIDDLEWARES, {optional: true})
+  protected expressMiddlewares: ExpressRequestHandler[] = [];
 
   constructor(
     // sonarignore:start
@@ -92,6 +95,15 @@ export class SecureSequence implements SequenceHandler {
         Remote Address = ${request.connection.remoteAddress}
         Remote Address (Proxy) = ${request.headers['x-forwarded-for']}`,
       );
+
+      if (this.expressMiddlewares?.length) {
+        const responseGenerated = await this.invokeMiddleware(
+          context,
+          this.expressMiddlewares,
+        );
+        if (responseGenerated) return;
+      }
+
       const finished = await this.invokeMiddleware(context);
       if (finished) return;
       const route = this.findRoute(request);
@@ -108,6 +120,7 @@ export class SecureSequence implements SequenceHandler {
       await this.authenticateClientRequest(request);
       const authUser: IAuthUserWithPermissions = await this.authenticateRequest(
         request,
+        response,
       );
       const isAccessAllowed: boolean = await this.checkAuthorisation(
         authUser?.permissions,