+++ packages/core/src/service-sequence.ts
@@ -1,5 +1,6 @@
 import {inject} from '@loopback/context';
 import {
+  ExpressRequestHandler,
   FindRoute,
   HttpErrors,
   InvokeMethod,
@@ -43,6 +44,8 @@ export class ServiceSequence implements SequenceHandler {
    */
   @inject(SequenceActions.INVOKE_MIDDLEWARE, {optional: true})
   protected invokeMiddleware: InvokeMiddleware = () => false;
+  @inject(SFCoreBindings.EXPRESS_MIDDLEWARES, {optional: true})
+  protected expressMiddlewares: ExpressRequestHandler[] = [];
 
   constructor(
     // sonarignore:start
@@ -77,6 +80,15 @@ export class ServiceSequence implements SequenceHandler {
         Remote Address = ${request.connection.remoteAddress}
         Remote Address (Proxy) = ${request.headers['x-forwarded-for']}`,
       );
+
+      if (this.expressMiddlewares?.length) {
+        const responseGenerated = await this.invokeMiddleware(
+          context,
+          this.expressMiddlewares,
+        );
+        if (responseGenerated) return;
+      }
+
       const finished = await this.invokeMiddleware(context);
       if (finished) return;
       const route = this.findRoute(request);
@@ -84,6 +96,7 @@ export class ServiceSequence implements SequenceHandler {
 
       const authUser: IAuthUserWithPermissions = await this.authenticateRequest(
         request,
+        response,
       );
       const isAccessAllowed: boolean = await this.checkAuthorisation(
         authUser?.permissions,