+++ packages/core/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-alpha.26](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/core@1.0.0-alpha.25...@sourceloop/core@1.0.0-alpha.26) (2021-02-06)
+
+
+### Bug Fixes
+
+* **core:** fixes in auth action in sequence ([#119](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/119)) ([1102072](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/110207228348bc901d6286f2cf17bafd8671baf6))
+
+
+
+
+
+# [1.0.0-alpha.25](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/core@1.0.0-alpha.24...@sourceloop/core@1.0.0-alpha.25) (2021-02-04)
+
+
+### Features
+
+* **core:** added swagger stats ([#118](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/118)) ([4413c2a](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/4413c2a5e400d834964d8cee2d4a3dae87b358b6))
+
+
+
+
+
+# [1.0.0-alpha.24](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/core@1.0.0-alpha.23...@sourceloop/core@1.0.0-alpha.24) (2021-02-01)
+
+
+### Bug Fixes
+
+* **core:** change in core to rectify publish issue ([#116](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/116)) ([194c5e8](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/194c5e8f4c33f2d088e84e0df8420fefa6ecd825))
+
+
+
+
+
+# [1.0.0-alpha.23](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/core@1.0.0-alpha.22...@sourceloop/core@1.0.0-alpha.23) (2021-01-19)
+
+
+### Features
+
+* **core:** SFO-169: added metrics functionality in core ([#114](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/114)) ([bc8ba4d](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/bc8ba4d1c429ee69279c3da3d854aab25484b695))
+
+
+
+
+
 # [1.0.0-alpha.22](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/core@1.0.0-alpha.21...@sourceloop/core@1.0.0-alpha.22) (2020-12-30)