+++ sandbox/auth-multitenant-example/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.17](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/auth-multitenant-example@2.0.0-alpha.16...@sourceloop/auth-multitenant-example@2.0.0-alpha.17) (2021-02-10)
+
+
+### Features
+
+* **authentication-service:** added express middleware in sequence ([#120](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/120)) ([bf81cb1](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/bf81cb12bb2b7cdd7edfb43dcf09c71abb891c1d))
+
+
+
+
+
+# [2.0.0-alpha.16](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/auth-multitenant-example@2.0.0-alpha.15...@sourceloop/auth-multitenant-example@2.0.0-alpha.16) (2021-02-06)
+
+
+### Bug Fixes
+
+* **core:** fixes in auth action in sequence ([#119](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/119)) ([1102072](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/110207228348bc901d6286f2cf17bafd8671baf6))
+
+
+
+
+
+# [2.0.0-alpha.15](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/auth-multitenant-example@2.0.0-alpha.14...@sourceloop/auth-multitenant-example@2.0.0-alpha.15) (2021-02-04)
+
+
+### Features
+
+* **core:** added swagger stats ([#118](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/118)) ([4413c2a](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/4413c2a5e400d834964d8cee2d4a3dae87b358b6))
+
+
+
+
+
+# [2.0.0-alpha.14](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/auth-multitenant-example@2.0.0-alpha.13...@sourceloop/auth-multitenant-example@2.0.0-alpha.14) (2021-02-01)
+
+**Note:** Version bump only for package @sourceloop/auth-multitenant-example
+
+
+
+
+
+# [2.0.0-alpha.13](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/auth-multitenant-example@2.0.0-alpha.12...@sourceloop/auth-multitenant-example@2.0.0-alpha.13) (2021-01-19)
+
+**Note:** Version bump only for package @sourceloop/auth-multitenant-example
+
+
+
+
+
 # [2.0.0-alpha.12](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/auth-multitenant-example@2.0.0-alpha.11...@sourceloop/auth-multitenant-example@2.0.0-alpha.12) (2020-12-30)
 
 **Note:** Version bump only for package @sourceloop/auth-multitenant-example