+++ services/notification-service/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-alpha.34](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/notification-service@1.0.0-alpha.33...@sourceloop/notification-service@1.0.0-alpha.34) (2021-02-06)
+
+**Note:** Version bump only for package @sourceloop/notification-service
+
+
+
+
+
+# [1.0.0-alpha.33](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/notification-service@1.0.0-alpha.32...@sourceloop/notification-service@1.0.0-alpha.33) (2021-02-04)
+
+
+### Bug Fixes
+
+* **notification-service:** hard deleting notif users for gdpr compliance ([#117](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/117)) ([af33640](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/af33640203a8c7029b7e0b52a44361c1f630ba0a))
+
+
+
+
+
+# [1.0.0-alpha.32](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/notification-service@1.0.0-alpha.31...@sourceloop/notification-service@1.0.0-alpha.32) (2021-02-01)
+
+**Note:** Version bump only for package @sourceloop/notification-service
+
+
+
+
+
+# [1.0.0-alpha.31](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/notification-service@1.0.0-alpha.30...@sourceloop/notification-service@1.0.0-alpha.31) (2021-01-30)
+
+
+### Features
+
+* **notification-service:** adds delete APIs for notification and notification users ([#115](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/115)) ([2462bd6](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/2462bd6f214726224257a1382f75df470db623c1))
+
+
+
+
+
+# [1.0.0-alpha.30](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/notification-service@1.0.0-alpha.29...@sourceloop/notification-service@1.0.0-alpha.30) (2021-01-19)
+
+**Note:** Version bump only for package @sourceloop/notification-service
+
+
+
+
+
 # [1.0.0-alpha.29](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/notification-service@1.0.0-alpha.28...@sourceloop/notification-service@1.0.0-alpha.29) (2020-12-30)
 
 **Note:** Version bump only for package @sourceloop/notification-service