+++ services/scheduler-service/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-alpha.34](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/scheduler-service@1.0.0-alpha.33...@sourceloop/scheduler-service@1.0.0-alpha.34) (2021-02-06)
+
+**Note:** Version bump only for package @sourceloop/scheduler-service
+
+
+
+
+
+# [1.0.0-alpha.33](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/scheduler-service@1.0.0-alpha.32...@sourceloop/scheduler-service@1.0.0-alpha.33) (2021-02-04)
+
+**Note:** Version bump only for package @sourceloop/scheduler-service
+
+
+
+
+
+# [1.0.0-alpha.32](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/scheduler-service@1.0.0-alpha.31...@sourceloop/scheduler-service@1.0.0-alpha.32) (2021-02-01)
+
+**Note:** Version bump only for package @sourceloop/scheduler-service
+
+
+
+
+
+# [1.0.0-alpha.31](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/scheduler-service@1.0.0-alpha.30...@sourceloop/scheduler-service@1.0.0-alpha.31) (2021-01-19)
+
+**Note:** Version bump only for package @sourceloop/scheduler-service
+
+
+
+
+
 # [1.0.0-alpha.30](https://github.com/sourcefuse/loopback4-microservice-catalog/compare/@sourceloop/scheduler-service@1.0.0-alpha.29...@sourceloop/scheduler-service@1.0.0-alpha.30) (2020-12-30)
 
 **Note:** Version bump only for package @sourceloop/scheduler-service